Fixed problem in Negation wish calculation
In order to address the problem of ambiguity of negated wishes if the
wish should be inverted to -1 or +1 in case of a 0 wish before negation
it was required to reintroduce the getDirection function in all
activators. Furthermore, the Conditional interface was extended to
forward this direction information as well in a dict similar to wishes.
On that account the corresponding unit tests have been extended and
improved