Not resetting activation when activation gets below threshold
In order to avoid quickly toggling between two conflicting behaviours it
was necessary to reset the activation calculation only in the case of
vilolated preconditions. A corresponding flag is introduced and
forwarded in the related functions.
Furthermore, this commit included some logging configuration
improvements as well as the feature of not executing planning in case of
a planBias of 0.0